@@ -1,7 +1,7 @@
 import { NestFactory, Reflector } from '@nestjs/core';
 import { AppModule } from './app.module';
 import * as express from 'express';
-const AppDataSource = require('./database/database.module');
+import * as cors from 'cors';
 
 console.log(
   `Application created, connecting to database ${process.env.DATABASEIP}`,
@@ -14,12 +14,9 @@ async function bootstrap() {
     origin: '*',
   };
 
-  const cors = require('cors');
-
   app.enableCors(corsOptions);
   app.use(cors(corsOptions));
 
-  if (AppDataSource.isInitialized === false) await AppDataSource.initialize();
   await app.init();
   await app.listen(3000);
 }
