@@ -4,24 +4,21 @@ import { WordModule } from './word/word.module';
 import { UserModule } from './user/user.module';
 import { PotentialWordModule } from './potential_word/potential_word.module';
 import { ThemeModule } from './theme/theme.module';
-import { AuthService } from './auth/auth.service';
-import { AuthModule } from './auth/auth.module';
 import { JwtStrategy } from './auth/jwt.strategy';
 import { JwtModule } from '@nestjs/jwt';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { jwtConstants } from './auth/jwt.constant';
 
 @Module({
   imports: [
-    JwtModule.register({
-      secret: process.env.JWT_SECRET,
-      signOptions: { expiresIn: '1d' },
-    }),
     DatabaseModule,
+    JwtModule.register({
+        secret: jwtConstants.secret,
+    }),
     WordModule,
     UserModule,
     PotentialWordModule,
     ThemeModule,
-    AuthModule,
-  ],
-  providers: [AuthService, JwtStrategy],
+  ]
 })
 export class AppModule {}
