@@ -1,4 +1,6 @@
+
+
 export const jwtConstants = {
-  secret: process.env.JWT_SECRET,
+  secret: "" + process.env.JWT_SECRET,
   expiresIn: '1d',
 };
