@@ -3,10 +3,11 @@ import { ThemeController } from './theme.controller';
 import { ThemeService } from './theme.service';
 import { Theme } from './theme.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { JwtService } from '@nestjs/jwt';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Theme])],
   controllers: [ThemeController],
-  providers: [ThemeService],
+  providers: [ThemeService, JwtService],
 })
 export class ThemeModule {}
